Update AUIScreenController.cs
Properties need to set the values when the UI is active, not when the UI is deactive. As when the properties are set in the UI there are a lot of stuff in UI which needs to be populated like dynamic list creation (might involve gameObject instantiation or deactivation) on the basis of the data passed.